Bosszuallo osztaly

+++ src/com/bzsomi/Szuperhos.java
@@ -3,6 +3,6 @@ package com.bzsomi;
 public interface Szuperhos {
 
     public boolean legyoziE(Szuperhos hos);
-    public  int  mekkoraAzEreje();
+    public  double  mekkoraAzEreje();
 
 }
